requirements.txt
fastapi==0.111.0
tinydb==4.8.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic~=2.10.6